adding support for bls header seal verification
this is still a wip as we depend on a local version of bls-crypto, the
library written by celo-org.
As it is, even though the library works and tests are passing, it does
not compile for wasm.

Further investigations needed
