+++ light_client/Cargo.toml
@@ -8,20 +8,24 @@ edition = "2018"
 [dependencies]
 serde = { version = "1", features = ["derive"]}
 celo-ibc = {path = "../ibc"}
-celo-types = {path = "../types"}
+celo-types = {path = "../types", features = ["bls-support"]}
 anomaly = { version = "0.2", default-features = false }
 thiserror = { version = "1.0", default-features = false }
 rlp = {version = "0.5"}
 base64 = { version = "0.13"}
 
+
 # cosmwasm contract deps
 cosmwasm-std = { version = "0.16", default-features = false }
 schemars = { version = "0.8"}
 
+#indirect dependency with wasm issues
+getrandom = {version = "0.2", features = ["js"]}
 
 [dev-dependencies]
 cosmwasm-vm = {version = "0.16", default-features = false }
 serde_json = {version = "1.0"}
+celo-types = {path = "../types", features = ["serialize"]}
 
 [lib]
 crate-type = ["cdylib", "rlib"]
