+++ types/go.mod
@@ -1,0 +1,38 @@
+module cmd/main.go
+
+go 1.17
+
+require github.com/celo-org/celo-blockchain v1.4.0
+
+require (
+	filippo.io/edwards25519 v1.0.0-alpha.2 // indirect
+	github.com/StackExchange/wmi v1.2.1 // indirect
+	github.com/VictoriaMetrics/fastcache v1.5.7 // indirect
+	github.com/aristanetworks/goarista v0.0.0-20170210015632-ea17b1a17847 // indirect
+	github.com/btcsuite/btcd v0.20.1-beta // indirect
+	github.com/celo-org/celo-bls-go v0.2.4 // indirect
+	github.com/cespare/xxhash/v2 v2.1.1 // indirect
+	github.com/deckarep/golang-set v0.0.0-20180603214616-504e848d77ea // indirect
+	github.com/go-ole/go-ole v1.2.5 // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/golang/snappy v0.0.3 // indirect
+	github.com/gorilla/websocket v1.4.1-0.20190629185528-ae1634f6a989 // indirect
+	github.com/hashicorp/golang-lru v0.5.4 // indirect
+	github.com/hdevalence/ed25519consensus v0.0.0-20201207055737-7fde80a9d5ff // indirect
+	github.com/holiman/uint256 v1.1.1 // indirect
+	github.com/huin/goupnp v1.0.0 // indirect
+	github.com/jackpal/go-nat-pmp v1.0.2-0.20160603034137-1fa385a6f458 // indirect
+	github.com/mattn/go-runewidth v0.0.4 // indirect
+	github.com/olekukonko/tablewriter v0.0.2-0.20190409134802-7e037d187b0c // indirect
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/prometheus/tsdb v0.6.2-0.20190402121629-4f204dcbc150 // indirect
+	github.com/shirou/gopsutil v2.20.5+incompatible // indirect
+	github.com/steakknife/bloomfilter v0.0.0-20180922174646-6819c0d2a570 // indirect
+	github.com/steakknife/hamming v0.0.0-20180906055917-c99c65617cd3 // indirect
+	github.com/syndtr/goleveldb v1.0.1-0.20190923125748-758128399b1d // indirect
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
+	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
+	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4 // indirect
+	golang.org/x/text v0.3.3 // indirect
+	gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce // indirect
+)
